restructure maverick to be used as both a package and a program
