+++ test/e2e/maverick/node/privval.go
@@ -1,4 +1,4 @@
-package main
+package node
 
 import (
 	"errors"
